CurrentModule = PlasmaSpecies

Index

PlasmaSpecies.DiNitrogen

PlasmaSpecies.Electron

PlasmaSpecies.Nitrogen

PlasmaSpecies.Species

PlasmaSpecies.Species

PlasmaSpecies.TetraNitrogen

PlasmaSpecies.TriNitrogen

PlasmaSpecies.apply_tree

API

PlasmaSpecies.DiNitrogen — Type

DiNitrogen <: Gas

Molecular nitrogen. Registered label: N2

PlasmaSpecies.Electron — Type

Electron <: Gas

Atomic nitrogen. Registered label: e

PlasmaSpecies.Nitrogen — Type

Nitrogen <: Gas

Atomic nitrogen. Registered label: N

PlasmaSpecies.Species — Type

Species(
    gas <: Gas, 
    charge=Neutral(),  
    electronic_state=nothing, 
    vibrational_state=nothing, 
    rotational_state=nothing
)

Fields

gas <: Gas: Label of the parent gas, e.g. a struct like Nitrogen <: Gas or the general StringGas(str:String).

charge=Neutral(): Charge of the species, e.g. Neutral(), Positive(n), Negative(n)

electronic_state=nothing: Optional, label for the electronic state.

vibrational_state=nothing: Optional, label for the vibrational state. If defined, electronic_state cannot be nothing.

rotational_state=nothing: Optional, label for the rotational state. If defined, vibrational_state cannot be nothing.

PlasmaSpecies.Species — Method

Species(str::String)

Convenience constructor for the Species struct. It parses a string of the format defined by the LoKI-B Boltzmann solver and automatically fills in the fields of Species.

PlasmaSpecies.TetraNitrogen — Type

TetraNitrogen <: Gas

Tetra atomic nitrogen. Registered label: N4

PlasmaSpecies.TriNitrogen — Type

TriNitrogen <: Gas

Tri atomic nitrogen. Registered label: N3

PlasmaSpecies.apply_tree — Method

```julia 
apply_tree(t::SpeciesTree, reactions::PlasmaReaction)
```

Apply the species tree to a reaction and return an array with tuples containing the scaling factor and the new reaction. This means for each participating species it is checked, if it is a leaf of the tree. If not, a new reaction is created for each descendent species. If the non-leaf species is a product, the branching factor is one over the number of descendants (effectively assuming that the total reaction rate is distributed equally over all possible products).